+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - When looking at the map for a group you will see the group itself on the map plus all descendant groups.
 - __Group relationship invites/requests:__ Groups can request to join other groups or invite a group to join them, and these invites/requests can be canceled, accepted or rejected by moderators of the other group. This all happens from the new group settings page "Related Groups".
 - __Group Join Questions:__ groups can now have questions that must be answered when a person is requesting to join the group. These are set up in the group settings page, and the form with the questions to answer shows up anywhere we show a Request to Join button to a user (when they are looking at a group they are not a member of that they are allowed to join because the accessibility setting is Restricted.)
-- New user settings page for Invites & Requests where you can see any current invites and requests to join a group and cancel, accept or reject them.
+- __Manage Invites & Requests__: New user settings page for Invites & Requests where you can see any current invites and requests to join a group and cancel, accept or reject them.
 - __Inline Comments:__ Comments can be nested undearneath other comments.
 - __Create Button:__ New button in the group navigation menu to create a post or a group within the current group.
 - __Tour__: A tour for new users to introduce the basics of Hylo when you first join